[RND-1239] NES 기반 oss 버전을 octopus로 업그레이드 (#35)
* try to upgrade oss to octopus

* reset .gitmodules temporarily

* link src/fmt and src/spawn

* apply octopus dashboard

* revert mgr/dashboard to nautilus version

* change commit id of src/fmt

* change commit id of src/googletest

* fix build errors

* remove unsupported mgr modules

* remove unsupported mgr modules

* add atlassian_statupage, objstore_usersync modules to mgr-modules-core rpm

* change release name to 'belfry'

* change release name to 'belfry'

* add belfry rpm

* change requirement of mgr-dashboard for centos7

* change python requirement for centos7

* remove duplicated option

* remove duplicated option

* revive python2 build

* add python2-compat to python codes

* add python2-compat to python codes2

* add python2-compat to ceph-mgr codes

* add python2-compat to ceph-mgr codes2

* add PYTHON_VERSION definition to src/ceph-volume/CMakeLists.txt

* add PYTHON_VERSION definition to src/python-common/CMakeLists.txt

* adjust rpm requisition

* correct python sys_path of ceph-mgr

* correct python sys_path of ceph-mgr

* add python2-compat to python codes3

* add python2-compat to python codes4(remove TypeVar and NewType)

* make python3 packages not to remove python2 packages

* add python2-compat to python codes4(remove Generic)

* add python2-compat to python codes6(remove class tamplate)

* add python2-compat to ceph-mgr codes3

* add python2-compat to python codes7(etc.)

* add python2-compat to ceph-mgr codes4

* remove redundunt code in src/mgr/ActivePyModule.cc

* make mgr-moudles to be compatable to python2.7 (crash, dashboard, objstore_usersync)

* fix circular import in ceph-volume

* src/pybind/mgr/mgr_module.py: add default 'inbuf' value of send_command()

* src/pybind/mgr/mgr_module.py: add _parse_opt_value()

* add missed import to src/ceph-volume/ceph_volume/util/system.py

* mgr/dashboard: add str_to_bool for ssl option

* Revert "mgr/dashboard: add str_to_bool for ssl option"

This reverts commit ca0dc5566f0a6f603d2e0eb5cadd80da2c8a802b.

* src/pybind/mgr/mgr_module.py: add missed _parse_opt_value()

* mgr/objstore_usersync: remove 'default' user collision, disable service of removing user in all ranger

* rgw_main.cc: remove critical misplaced 'destroy_global_ranger_manager()'

* rgw/rgw_ranger.cc: enhence ranger_manager destroy routine

* rgw/rgw_ranger_native.cc: remove redundunt logs

* add PYTHON_EXECUTABLE variable to CMakeList files

* change python version of ceph & ceph-crash

* reorder find_package for python in CMakeList file

+++ src/tools/rbd_mirror/image_deleter/TrashMoveRequest.cc
@@ -13,6 +13,8 @@
 #include "librbd/TrashWatcher.h"
 #include "librbd/Utils.h"
 #include "librbd/journal/ResetRequest.h"
+#include "librbd/mirror/ImageRemoveRequest.h"
+#include "librbd/mirror/GetInfoRequest.h"
 #include "librbd/trash/MoveRequest.h"
 #include "tools/rbd_mirror/image_deleter/Types.h"
 
@@ -69,39 +71,45 @@ void TrashMoveRequest<I>::handle_get_mirror_image_id(int r) {
     return;
   }
 
-  get_tag_owner();
+  get_mirror_info();
 }
 
 template <typename I>
-void TrashMoveRequest<I>::get_tag_owner() {
+void TrashMoveRequest<I>::get_mirror_info() {
   dout(10) << dendl;
 
   auto ctx = create_context_callback<
-    TrashMoveRequest<I>, &TrashMoveRequest<I>::handle_get_tag_owner>(this);
-  librbd::Journal<I>::get_tag_owner(m_io_ctx, m_image_id, &m_mirror_uuid,
-                                    m_op_work_queue, ctx);
+    TrashMoveRequest<I>, &TrashMoveRequest<I>::handle_get_mirror_info>(this);
+  auto req = librbd::mirror::GetInfoRequest<I>::create(
+    m_io_ctx, m_op_work_queue, m_image_id, &m_mirror_image, &m_promotion_state,
+    &m_primary_mirror_uuid, ctx);
+  req->send();
 }
 
 template <typename I>
-void TrashMoveRequest<I>::handle_get_tag_owner(int r) {
+void TrashMoveRequest<I>::handle_get_mirror_info(int r) {
   dout(10) << "r=" << r << dendl;
 
-  if (r < 0 && r != -ENOENT) {
+  if (r == -ENOENT) {
+    dout(5) << "image " << m_global_image_id << " is not mirrored" << dendl;
+    finish(r);
+    return;
+  } else if (r < 0) {
     derr << "error retrieving image primary info for image "
          << m_global_image_id << ": " << cpp_strerror(r) << dendl;
     finish(r);
     return;
-  } else if (r != -ENOENT) {
-    if (m_mirror_uuid == librbd::Journal<>::LOCAL_MIRROR_UUID) {
-      dout(10) << "image " << m_global_image_id << " is local primary" << dendl;
-      finish(-EPERM);
-      return;
-    } else if (m_mirror_uuid == librbd::Journal<>::ORPHAN_MIRROR_UUID &&
-               !m_resync) {
-      dout(10) << "image " << m_global_image_id << " is orphaned" << dendl;
-      finish(-EPERM);
-      return;
-    }
+  }
+
+  if (m_promotion_state == librbd::mirror::PROMOTION_STATE_PRIMARY) {
+    dout(10) << "image " << m_global_image_id << " is local primary" << dendl;
+    finish(-EPERM);
+    return;
+  } else if (m_promotion_state == librbd::mirror::PROMOTION_STATE_ORPHAN &&
+             !m_resync) {
+    dout(10) << "image " << m_global_image_id << " is orphaned" << dendl;
+    finish(-EPERM);
+    return;
   }
 
   disable_mirror_image();
@@ -111,12 +119,10 @@ template <typename I>
 void TrashMoveRequest<I>::disable_mirror_image() {
   dout(10) << dendl;
 
-  cls::rbd::MirrorImage mirror_image;
-  mirror_image.global_image_id = m_global_image_id;
-  mirror_image.state = cls::rbd::MIRROR_IMAGE_STATE_DISABLING;
+  m_mirror_image.state = cls::rbd::MIRROR_IMAGE_STATE_DISABLING;
 
   librados::ObjectWriteOperation op;
-  librbd::cls_client::mirror_image_set(&op, m_image_id, mirror_image);
+  librbd::cls_client::mirror_image_set(&op, m_image_id, m_mirror_image);
 
   auto aio_comp = create_rados_callback<
     TrashMoveRequest<I>,
@@ -147,11 +153,66 @@ void TrashMoveRequest<I>::handle_disable_mirror_image(int r) {
     return;
   }
 
+  open_image();
+}
+
+template <typename I>
+void TrashMoveRequest<I>::open_image() {
+  dout(10) << dendl;
+
+  m_image_ctx = I::create("", m_image_id, nullptr, m_io_ctx, false);
+
+  // ensure non-primary images can be modified
+  m_image_ctx->read_only_mask &= ~librbd::IMAGE_READ_ONLY_FLAG_NON_PRIMARY;
+
+  {
+    // don't attempt to open the journal
+    std::unique_lock image_locker{m_image_ctx->image_lock};
+    m_image_ctx->set_journal_policy(new JournalPolicy());
+  }
+
+  Context *ctx = create_context_callback<
+    TrashMoveRequest<I>, &TrashMoveRequest<I>::handle_open_image>(this);
+  m_image_ctx->state->open(librbd::OPEN_FLAG_SKIP_OPEN_PARENT, ctx);
+}
+
+template <typename I>
+void TrashMoveRequest<I>::handle_open_image(int r) {
+  dout(10) << "r=" << r << dendl;
+
+  if (r == -ENOENT) {
+    dout(5) << "mirror image does not exist, removing orphaned metadata" << dendl;
+    m_image_ctx = nullptr;
+    remove_mirror_image();
+    return;
+  }
+
+  if (r < 0) {
+    derr << "failed to open image: " << cpp_strerror(r) << dendl;
+    m_image_ctx->destroy();
+    m_image_ctx = nullptr;
+    finish(r);
+    return;
+  }
+
+  if (m_image_ctx->old_format) {
+    derr << "cannot move v1 image to trash" << dendl;
+    m_ret_val = -EINVAL;
+    close_image();
+    return;
+  }
+
   reset_journal();
 }
 
 template <typename I>
 void TrashMoveRequest<I>::reset_journal() {
+  if (m_mirror_image.mode == cls::rbd::MIRROR_IMAGE_MODE_SNAPSHOT) {
+    // snapshot-based mirroring doesn't require journal feature
+    acquire_lock();
+    return;
+  }
+
   dout(10) << dendl;
 
   // ensure that if the image is recovered any peers will split-brain
@@ -169,45 +230,7 @@ void TrashMoveRequest<I>::handle_reset_journal(int r) {
 
   if (r < 0 && r != -ENOENT) {
     derr << "failed to reset journal: " << cpp_strerror(r) << dendl;
-    finish(r);
-    return;
-  }
-
-  open_image();
-}
-
-template <typename I>
-void TrashMoveRequest<I>::open_image() {
-  dout(10) << dendl;
-
-  m_image_ctx = I::create("", m_image_id, nullptr, m_io_ctx, false);
-
-  {
-    // don't attempt to open the journal
-    RWLock::WLocker snap_locker(m_image_ctx->snap_lock);
-    m_image_ctx->set_journal_policy(new JournalPolicy());
-  }
-
-  Context *ctx = create_context_callback<
-    TrashMoveRequest<I>, &TrashMoveRequest<I>::handle_open_image>(this);
-  m_image_ctx->state->open(librbd::OPEN_FLAG_SKIP_OPEN_PARENT, ctx);
-}
-
-template <typename I>
-void TrashMoveRequest<I>::handle_open_image(int r) {
-  dout(10) << "r=" << r << dendl;
-
-  if (r < 0) {
-    derr << "failed to open image: " << cpp_strerror(r) << dendl;
-    m_image_ctx->destroy();
-    m_image_ctx = nullptr;
-    finish(r);
-    return;
-  }
-
-  if (m_image_ctx->old_format) {
-    derr << "cannot move v1 image to trash" << dendl;
-    m_ret_val = -EINVAL;
+    m_ret_val = r;
     close_image();
     return;
   }
@@ -217,12 +240,18 @@ void TrashMoveRequest<I>::handle_open_image(int r) {
 
 template <typename I>
 void TrashMoveRequest<I>::acquire_lock() {
-  m_image_ctx->owner_lock.get_read();
+  m_image_ctx->owner_lock.lock_shared();
   if (m_image_ctx->exclusive_lock == nullptr) {
-    derr << "exclusive lock feature not enabled" << dendl;
-    m_image_ctx->owner_lock.put_read();
-    m_ret_val = -EINVAL;
-    close_image();
+    m_image_ctx->owner_lock.unlock_shared();
+
+    if (m_mirror_image.mode == cls::rbd::MIRROR_IMAGE_MODE_SNAPSHOT) {
+      // snapshot-based mirroring doesn't require exclusive-lock
+      trash_move();
+    } else {
+      derr << "exclusive lock feature not enabled" << dendl;
+      m_ret_val = -EINVAL;
+      close_image();
+    }
     return;
   }
 
@@ -232,7 +261,7 @@ void TrashMoveRequest<I>::acquire_lock() {
     TrashMoveRequest<I>, &TrashMoveRequest<I>::handle_acquire_lock>(this);
   m_image_ctx->exclusive_lock->block_requests(0);
   m_image_ctx->exclusive_lock->acquire_lock(ctx);
-  m_image_ctx->owner_lock.put_read();
+  m_image_ctx->owner_lock.unlock_shared();
 }
 
 template <typename I>
@@ -288,15 +317,12 @@ template <typename I>
 void TrashMoveRequest<I>::remove_mirror_image() {
   dout(10) << dendl;
 
-  librados::ObjectWriteOperation op;
-  librbd::cls_client::mirror_image_remove(&op, m_image_id);
-
-  auto aio_comp = create_rados_callback<
+  auto ctx = create_context_callback<
     TrashMoveRequest<I>,
     &TrashMoveRequest<I>::handle_remove_mirror_image>(this);
-  int r = m_io_ctx.aio_operate(RBD_MIRRORING, aio_comp, &op);
-  ceph_assert(r == 0);
-  aio_comp->release();
+  auto req = librbd::mirror::ImageRemoveRequest<I>::create(
+    m_io_ctx, m_global_image_id, m_image_id, ctx);
+  req->send();
 }
 
 template <typename I>
@@ -318,6 +344,10 @@ template <typename I>
 void TrashMoveRequest<I>::close_image() {
   dout(10) << dendl;
 
+  if (m_image_ctx == nullptr) {
+    handle_close_image(0);
+    return;
+  }
   Context *ctx = create_context_callback<
     TrashMoveRequest<I>, &TrashMoveRequest<I>::handle_close_image>(this);
   m_image_ctx->state->close(ctx);
@@ -327,8 +357,10 @@ template <typename I>
 void TrashMoveRequest<I>::handle_close_image(int r) {
   dout(10) << "r=" << r << dendl;
 
-  m_image_ctx->destroy();
-  m_image_ctx = nullptr;
+  if (m_image_ctx != nullptr) {
+    m_image_ctx->destroy();
+    m_image_ctx = nullptr;
+  }
 
   if (r < 0) {
     derr << "failed to close image: " << cpp_strerror(r) << dendl;
